deprecate Page#markdown
Page#markdown used to be a reference to the markdown source of the page.
There should be need for a page object to access its own original
markdown source. For the time being, it is delegated to _markdown.source

Presently, only emitting a logging error (does not stop the build or
require --force)

Next step: promote from logging error to grunt.fail.warn (which requires
--force)

Finally: remove the deprecated accessor altogether